Alleviates some issues w/ cpp div

+++ std/cpp/chrono.hpp
@@ -8,6 +8,6 @@ auto time()
 
 std::ostream& operator <<(std::ostream& os, const std::chrono::duration<long int, std::ratio<1l, 1000000000l>>& t)
 {
-    os << std::chrono::duration_cast<std::chrono::duration<float>>(t).count();
+    os << std::chrono::duration_cast<std::chrono::duration<double>>(t).count();
     return os;
 }
